// src/app/model/student.ts
export interface Student {
    id: number;
    studentId: string;
    name: string;
    birthdate: string;  
    address: string;
    phone_number: string;
    email: string;
    password: string;
    dateOfJoining: string;  
    yearOfStudy: number;
    isYearCompleted: boolean;
    department: string;
    courseIds: number[];
  }
  